+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -149,7 +149,6 @@
 #import "SentryTraceContext.h"
 #import "SentryTracer+Test.h"
 #import "SentryTracer.h"
-#import "SentryTransaction+Private.h"
 #import "SentryTransaction.h"
 #import "SentryTransactionContext+Private.h"
 #import "SentryTransport.h"